+++ src/ConfBase.h
@@ -51,7 +51,9 @@ public:
       backupParameters(parent->backupParameters),
       prunePolicy(parent->prunePolicy),
       pruneParameters(parent->pruneParameters), preVolume(parent->preVolume),
-      postVolume(parent->postVolume), rsyncTimeout(parent->rsyncTimeout),
+      postVolume(parent->postVolume),
+      backupJobTimeout(parent->backupJobTimeout),
+      rsyncIOTimeout(parent->rsyncIOTimeout),
       rsyncCommand(parent->rsyncCommand),
       rsyncBaseOptions(parent->rsyncBaseOptions),
       rsyncExtraOptions(parent->rsyncExtraOptions),
@@ -84,8 +86,11 @@ public:
   /** @brief Post-volume hook */
   std::vector<std::string> postVolume;
 
-  /** @brief rsync timeout */
-  int rsyncTimeout = 0;
+  /** @brief backup job timeout */
+  int backupJobTimeout = 0;
+
+  /** @brief rsync IO timeout */
+  int rsyncIOTimeout = 0;
 
   /** @brief rsync command */
   std::string rsyncCommand = "rsync";